Clean up the functional test
